Rename common.rs to base.rs
